requirements.txt
yfinance>=0.2.31
pandas>=2.0.0
plotly>=5.18.0
typer[all]>=0.9.0
rich>=13.7.0
questionary>=2.0.1
python-dateutil>=2.8.2 